Call Path Error Reporting
One problem we had after Purging Macros
189f80b84b9a2e6779c4562ca57060cbd0bae7b9 is not having any sort of
stack trace when a type error occurs somewhere. In case of the macros
we had at least the expansion stack that was tracked by the tokens.

For procs we are using the ret_stack from the type checking context
now. It implicitly stores the current call path which can be used to
generate useful compiler diagnostics.