Type Checking Call Limit
Right now our type checking process hangs on recursive functions. It
is useful to have recursion in a programming language, but it's not
critical for making Porth self-hosted as of right now. So for now if
the type checking process makes too many function calls we just fail
the compilation process suggesting a workaround (disabling the type
checking by providing the `-unsafe` flag).

We gonna implement proper type checking of the recursive function
calls a bit later.