[porth.py] introduce `check` subcommand
It simply compiles the program to IR and type checks it. No code
generation. No execution. Just a simple sanity check.

I introduced it because as of right now the code generation step is
pretty expensive (thanks to nasm (we will fix it later)). And when I
work on a porth program I don't usually need to run it right
away. Only type check it as I incrementally write it piece by piece to
ensure that everything's fine.