Fix stack underflow diagnostic in compile time evaluation
The diagnostic was using `Token.value` as the name of the "word" to
report, but `Token.value` is not always a string for all
`Token.type`-s. For `TOKEN_KEYWORD` for example it is an integer (the
id of the token). `Token.text` is the original text of the token from
the source code file.

Not an ideal solution, but the compile time evaluation will be
rewritten in the future anyway to enable majority of the language
features at compile time. So some of the existing type checking code
might be reused in here.